tommy 09/07/2021 11:57

+++ ft_itoa.c
@@ -45,7 +45,7 @@ static char	*ft_reverse (char *str)
 	return (aux);
 }
 
-char	*ft_itoa (int n)
+char	*ft_itoa (long int n)
 {
 	char	*a;
 	int		i;
